refactor: separate game class into module
